Updated select2 with selectWoo
